+++ tests/mmap-gettest.c
@@ -47,7 +47,7 @@ extern int errno;
 static char hugetlb_mount[BUF_SZ];
 static unsigned int iter;
 static int nr_hugepages;
-static size_t hpage_size;
+static long hpage_size;
 
 static int do_one(char *mountpoint, size_t size) {
 	char *ma;
